endpoint logic removed to UseCase class
